+++ lib/twilio-ruby/rest/api/v2010/account/message.rb
@@ -32,18 +32,19 @@ module Twilio
             # @param [String] to The destination phone number for SMS/MMS or a [Channel user
             #   address](https://www.twilio.com/docs/sms/channels#channel-addresses) for other
             #   3rd party channels. Destination phone numbers should be formatted with a '+' and
-            #   country code e.g., +16175551212
+            #   country code e.g., `+16175551212`
             #   ([E.164](https://www.twilio.com/docs/glossary/what-e164) format).
             # @param [String] status_callback A URL where Twilio will POST each time your
             #   message status changes to one of the following: `queued`, `failed`, `sent`,
-            #   `delivered`, or `undelivered`. Twilio will POST the `MessageSid`,
-            #   `MessageStatus`, and `ErrorCode` along with the other [standard request
-            #   parameters](https://www.twilio.com/docs/sms/twiml#request-parameters). If you
-            #   include this parameter in addition to a `MessagingServiceSid`, Twilio will
+            #   `delivered`, or `undelivered`. Twilio will POST its [standard request
+            #   parameters](https://www.twilio.com/docs/sms/twiml#request-parameters) as well as
+            #   some additional parameters including `MessageSid`, `MessageStatus`, and
+            #   `ErrorCode`(see more details [below](#statuscallback-request-parameters)). If
+            #   you include this parameter in addition to a `MessagingServiceSid`, Twilio will
             #   override the Status Callback URL of the [Messaging
             #   Service](https://www.twilio.com/docs/sms/services/api). URLs must contain a
             #   valid hostname – underscores are not allowed.
-            # @param [String] application_sid Twilio will POST `MessageSid` as well as
+            # @param [String] application_sid Twilio will POST a `MessageSid` as well as
             #   `MessageStatus=sent` or `MessageStatus=failed` to the URL in the
             #   `MessageStatusCallback` property of this
             #   [application](https://www.twilio.com/docs/usage/api/applications). If the
@@ -52,8 +53,8 @@ module Twilio
             # @param [String] max_price The total maximum price up to the fourth decimal
             #   (0.0001) in US dollars acceptable for the message to be delivered. *All messages
             #   will be queued for delivery regardless of the price point.* A POST request will
-            #   later be made to your Status Callback URL with a status change of 'Sent' or
-            #   'Failed'. When the price of the message is greater than this value, the message
+            #   later be made to your Status Callback URL with a status change of `Sent` or
+            #   `Failed`. When the price of the message is greater than this value, the message
             #   will fail and not be sent. When `MaxPrice` is not set, all prices for the
             #   message are accepted.
             # @param [Boolean] provide_feedback Set this value to `true` if you are sending
@@ -75,19 +76,26 @@ module Twilio
             # @param [Boolean] smart_encoded The smart_encoded
             # @param [String] from A Twilio phone number (in
             #   [E.164](https://www.twilio.com/docs/glossary/what-e164) format),  [alphanumeric
-            #   sender ID](https://www.twilio.com/docs/sms/send-messages#alphanumeric-sender-id)
+            #   sender
+            #   ID](https://www.twilio.com/docs/sms/send-messages#use-an-alphanumeric-sender-id)
             #   or a [Channel Endpoint
             #   address](https://www.twilio.com/docs/sms/channels#channel-addresses) enabled for
             #   the type of message you wish to send. Phone numbers or [short
             #   codes](https://www.twilio.com/docs/sms/api/short-codes) purchased from Twilio
             #   work here. You cannot (for example) spoof messages from your own cell phone
-            #   number. *Should not be passed if you are using `MessagingServiceSid`.*
+            #   number. *Do not use this parameter if you are using `MessagingServiceSid`.*
             # @param [String] messaging_service_sid The 34-character unique ID of the
             #   [Messaging
-            #   Service](https://www.twilio.com/docs/api/messaging/send-messages#messaging-services) you want to associate with this Message. Set this parameter to use the Messaging Service Settings and [Copilot Features](https://www.twilio.com/docs/sms/services) you have configured. When only this parameter is set, Twilio will use your enabled Copilot Features to select the `From` phone number for delivery. *Should not be passed if you are using `From`.*
+            #   Service](https://www.twilio.com/docs/sms/services#send-a-message-with-copilot)
+            #   you want to associate with this Message. Set this parameter to use the
+            #   [Messaging Service Settings and Copilot
+            #   Features](https://www.twilio.com/console/sms/services) you have configured. When
+            #   only this parameter is set, Twilio will use your enabled Copilot Features to
+            #   select the `From` phone number for delivery. *Do not pass this value if you are
+            #   using `From`.*
             # @param [String] body The text of the message you want to send, limited to 1600
             #   characters.
-            # @param [String] media_url The URL of the media you wish to send with the
+            # @param [String] media_url The URL containing the media you wish to send with the
             #   message. `gif` , `png` and `jpeg` content is currently supported and will be
             #   formatted correctly on the recipient's device. [Other
             #   types](https://www.twilio.com/docs/sms/accepted-mime-types) are also accepted by