+++ spec/integration/video/v1/recording_spec.rb
@@ -43,6 +43,7 @@ describe 'Recording' do
           "container_format": "mka",
           "codec": "OPUS",
           "track_name": "A name",
+          "offset": 10,
           "grouping_sids": {
               "room_sid": "RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
           },
@@ -119,6 +120,7 @@ describe 'Recording' do
                   "container_format": "mka",
                   "codec": "OPUS",
                   "track_name": "A name",
+                  "offset": 10,
                   "grouping_sids": {
                       "room_sid": "RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                       "participant_sid": "PAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"